Added minimap PNG.
Not skilled enough to know how to implement this yet as I've never used react or three.js, but plan to return to it if no one else gets to it first. This is a png of the Paris Prominade map for someone to use while implementing a mini map. Added to images folder.